JSSC-66  🔨 Refactor : CreateExample 파일 uploadpost폴더로 옮김
